requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
fastapi 
uvicorn 
pydantic 
python-multipart
